Basic zoning, automatic server swapping
Some moving around of files. Switch to stable instead of nightly.
Saving of a player's position. It isn't really feasible to rely on the
servers themselves, unless we want to just search through their world
files for player positions before we choose the server they need to
join. Saving happens every 30 seconds, and on shutdown. Auto swapping
uses the continuous feed of position information from the server side
plugin to figure out the zones that a player is in, and then figures
out what connections to disconnect, connect, and swap.